Fixed Inconsistencies and Bugs
The program was using the PORT set of bits to change the voltage on the pin.  This was incorrect.  It has been corrected to use the LAT bits now.

The pin was changed from B4 to B5 to avoid input-only pins

Various names were changed in comments to remain consistent with the change in pins.

In an unrelated note, physical changes were made to the PIC base configuration in order to allow the program to be written onto the PIC.