Reorganized the A/D Conversion Method and Changed Values for PWM
Added more code to the A/D Conversion method, which should complete it.  The only thing left to do regarding A/D Conversion is to sample the pin and let the conversion happen, and then to obtain and use the value.

The PWM values being changed are of no importance, they were simply changed to accomodate a different example.